Added new classes
Split up the PetGUI class so that I had a more OOP design of the GUI. Much easier to access code and look for related code
